DB-001: billing_payments.amount_eur is Numeric(14, 2), not Float
The only float money column in the schema — every other amount is
Numeric(14, 2). It is the server-side record a redirect-flow (EveryPay)
payment result is VERIFIED against: a float 29.99 reloads as
29.989999999999998, so an exact comparison with the provider's "29.99"
fails on a correct payment and any sum over the table drifts. The model,
the provider protocol, the route and the recurring-charge service now pass
Decimal end to end; migration e6a8c0d2f4b6 retypes the column with
round(amount_eur::numeric, 2) on Postgres (every value ever written came
from an integer plan price, so nothing is lost). Round-tripped on Postgres
16, alembic check clean. Tests: a 29.99 round-trips exactly as a Decimal;
checkout stores the plan price as Decimal("99.00"). README's revision
counter: 127.

+++ backend/app/api/routes/billing.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import logging
 import uuid
+from decimal import Decimal
 
 from fastapi import APIRouter, Depends, HTTPException, Request, status
 from fastapi.responses import RedirectResponse
@@ -175,7 +176,7 @@ async def start_checkout(
         session = await provider.start_checkout(
             org_id=org.id,
             plan_key=body.plan,
-            amount_eur=float(target.price_eur),
+            amount_eur=Decimal(target.price_eur),
             order_reference=order_reference,
             customer_id=customer_id,
         )
@@ -191,7 +192,7 @@ async def start_checkout(
                 reference=session.reference,
                 order_reference=order_reference,
                 plan_key=body.plan,
-                amount_eur=float(target.price_eur),
+                amount_eur=Decimal(target.price_eur),
                 state="initial",
             )
         )
